Added the test help functions in utils_test.go and changed the test code to work with them
